Cleanup of Configuration managers
- deleted duplicate ServerConfigurationManager
- Moved TestConfigurationManager from server to CLI package where all other config managers were already located
